Migrate the FlutterSampleNotificationProvider to the new EditorNotificationProvider API (#7840)
This is progress on https://github.com/flutter/flutter-intellij/issues/7830
